+++ Z/Data/Vector.hs
@@ -80,7 +80,7 @@ module Z.Data.Vector (
   , null
   , length
   , append
-  , map, map', imap', traverseVec, traverseWithIndex, traverseVec_, traverseWithIndex_
+  , map, map', imap', traverse, traverseWithIndex, traverse_, traverseWithIndex_
   , foldl', ifoldl', foldl1', foldl1Maybe'
   , foldr', ifoldr', foldr1', foldr1Maybe'
     -- ** Special folds
@@ -96,7 +96,7 @@ module Z.Data.Vector (
   , mapAccumR
   -- ** Generating and unfolding vector
   , replicate
-  , replicateMVec
+  , replicateM
   , cycleN
   , unfoldr
   , unfoldrN